fix: correctness bugs from full audit (display/focus, spawn/wait, virtio, fs, arch)
Audit of BeetOS-added code (display/focus lifecycle, spawn/wait, virtio
drivers, userspace fs, AArch64 arch) surfaced ~25 correctness issues; this
fixes all P0/P1/P2 and most P3. See audit-plan.md for the full list.

P0 — data corruption / system freeze (reproducible on QEMU):
- fs paths: reject over-long paths instead of silently truncating (rm/write
  were operating on the wrong file); pack_path_checked + shell guards.
- shell `cd -`: materialise target before overwriting PREV_BUF (self-aliasing
  corrupted the new CWD).
- shell `write`: derive content from raw args, not find() on the resolved path.
- kernel WaitProcess: exit-code reaper so a child that exits before the wait
  registers returns its real code, not a fabricated 0.
- display waiters: purge dying PID from the queue (PID reuse defeated the
  lazy is_err() skip → FB mapped into an innocent process); handle full queue
  (was suspending a thread that was never enqueued).
- userspace panic handlers: TerminateProcess instead of spinning (a panic
  while holding the display froze the whole console).

P1 — logic/API:
- ReleaseInputFocus now checks focus ownership; input focus cleared when its
  holder exits (dangling SID dropped keystrokes).
- AcquireDisplay idempotent for the current owner (was self-deadlock).
- WaitProcess rejects self / PID 1; wakeup no longer capped at 64 waiters.
- spawn_by_name_with_args frees the process slot on any setup failure.

P2 — latent on QEMU, fatal on real hardware:
- ELF load: I-cache sync (dc cvau / ic ivau) for executable pages.
- virtio pop_used: acquire barrier after the used-idx load (virtio_rmb); bound
  device-written descriptor id; free descriptor chain on blk timeout.
- mem: fix u32 overflow in MemoryRangeExtra address math.
- fb_console: use stride (not width) for scroll/clear.
- AArch64 syscall wrapper: loop on RetryCall (was dropping messages under
  server-queue backpressure).

P3 — robustness / project rules:
- W^X guard in the kernel PTE branch; D-cache clean on page invalidate.
- ELF header/phdr/reloc bounds validation.
- tarfs: octal parse handles leading spaces + base-256; list() path-boundary
  match; ls of a nonexistent disk dir now fails.
- misc: checked_add on lba+count, filter GIC INTIDs 1020-1022, ICMP reply
  bounded by IP total-length, graceful handling replacing panics.

Cross-build (aarch64-unknown-none) clean; 52 hosted kernel tests pass.

+++ xous/kernel/src/mem.rs
@@ -63,7 +63,11 @@ impl core::fmt::Display for MemoryRangeExtra {
             ((self.mem_tag >> 24) & 0xff) as u8 as char,
             self.mem_tag,
             self.mem_start,
-            self.mem_start + self.mem_size,
+            // Widen before adding: `mem_start + mem_size` in u32 wraps for a
+            // region ending at/above 4 GiB (debug builds panic).  The fields
+            // stay u32 because they are memcpy'd from the fixed 16-byte
+            // kernel-args layout; a 64-bit region format is an M3b task.
+            self.mem_start as u64 + self.mem_size as u64,
             self.mem_size
         )
     }
@@ -867,7 +871,7 @@ impl MemoryManager {
         offset += RAM_PAGES;
         // Go through additional regions looking for this address
         for region in &self.extra_regions {
-            if addr >= (region.mem_start as usize) && addr < (region.mem_start + region.mem_size) as usize {
+            if addr >= (region.mem_start as usize) && addr < (region.mem_start as usize) + (region.mem_size as usize) {
                 offset += (addr - (region.mem_start as usize)) / PAGE_SIZE;
                 return Some(offset);
             }
